+++ library/test/src/test/java/com/bumptech/glide/manager/DefaultConnectivityMonitorFactoryTest.java
@@ -31,42 +31,20 @@ public class DefaultConnectivityMonitorFactoryTest {
   }
 
   @Test
-  @Config(sdk = 18)
-  public void testDefaultConnectivityMonitorInjectPreMConnectivityStrategyImplWhenHasPermission() {
-    ShadowApplication.getInstance().grantPermissions("android.permission.ACCESS_NETWORK_STATE");
-    ConnectivityMonitor connectivityMonitor =
-        factory.build(
-            RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
-    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
-    DefaultConnectivityMonitor defaultConnectivityMonitor =
-        (DefaultConnectivityMonitor) connectivityMonitor;
-    assertThat(defaultConnectivityMonitor.strategy)
-        .isInstanceOf(PreMConnectivityStrategyImpl.class);
-  }
-
-  @Test
   @Config(sdk = 23)
-  public void testDefaultConnectivityMonitorInjectPostMConnectivityStrategyImplWhenHasPermission() {
+  public void testReturnPostMConnectivityMonitorWhenHasPermission() {
     ShadowApplication.getInstance().grantPermissions("android.permission.ACCESS_NETWORK_STATE");
     ConnectivityMonitor connectivityMonitor =
         factory.build(
             RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
-    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
-    DefaultConnectivityMonitor defaultConnectivityMonitor =
-        (DefaultConnectivityMonitor) connectivityMonitor;
-    assertThat(defaultConnectivityMonitor.strategy)
-        .isInstanceOf(PostMConnectivityStrategyImpl.class);
+    assertThat(connectivityMonitor).isInstanceOf(PostMConnectivityMonitor.class);
   }
 
   @Test
-  public void testDefaultConnectivityMonitorInjectNullConnectivityStrategyImplNotHavePermission() {
+  public void testReturnsNullConnectivityMonitorWhenDoesNotHavePermission() {
     ConnectivityMonitor connectivityMonitor =
         factory.build(
             RuntimeEnvironment.application, mock(ConnectivityMonitor.ConnectivityListener.class));
-    assertThat(connectivityMonitor).isInstanceOf(DefaultConnectivityMonitor.class);
-    DefaultConnectivityMonitor defaultConnectivityMonitor =
-        (DefaultConnectivityMonitor) connectivityMonitor;
-    assertThat(defaultConnectivityMonitor.strategy)
-        .isInstanceOf(NullConnectivityStrategyImpl.class);
+    assertThat(connectivityMonitor).isInstanceOf(NullConnectivityMonitor.class);
   }
 }
